Role: Coder-R195 extract reusable host-v2 donor adapter module for Gate D #252
Carry over the donor-side host-v2 adapter path into reusable production-side code and route CAST + hash join donor-native tests through that adapter path instead of fixture-local host wiring.

Refs zanmato1984/tiforth#252

Refs zanmato1984/tiforth#77
